app/requirements.txt
Django==3.1.2
gunicorn==20.0.4
psycopg2==2.8.6
djangorestframework==3.12.1
